TimeMon: fixed running on modern FreeBSD.

+++ Applications/TimeMon/loadave.c
@@ -8,7 +8,7 @@ int la_init(unsigned long long *times)
   return la_read(times);
 }
 
-#if defined( linux )
+#if defined(linux)
 
 int la_read(unsigned long long *times)
 {
@@ -32,36 +32,29 @@ int la_read(unsigned long long *times)
   return LA_NOERR;
 }
 
-#elif defined( __FreeBSD__ ) 
+#elif defined(__FreeBSD__)
 
 #include <sys/types.h>
 #include <sys/errno.h>
 #include <sys/resource.h>       // CPUSTATES
 #include <sys/sysctl.h>         // sysctlbyname()
-#include <string.h>
-          
+#include <string.h>             // strerror()
+
 int la_read(unsigned long long *times)
 {
-  const char
-    *name = "kern.cp_time";
-  int
-    cpu_states[CPUSTATES];
-  size_t
-    nlen = sizeof cpu_states,
-    len = nlen;
-  int
-    err;
-  
-  err= sysctlbyname(name, &cpu_states, &nlen, NULL, 0);
-  if( -1 == err )
-  {
+  const char *name = "kern.cp_time";
+  long cpu_states[CPUSTATES];
+  size_t nlen = sizeof(cpu_states), len = nlen;
+  int err;
+
+  err = sysctlbyname(name, &cpu_states, &nlen, NULL, 0);
+  if (-1 == err) {
     fprintf(stderr, "sysctl(%s...) failed: %s\n", name, strerror(errno));
     exit(errno);
   }
-  if( nlen != len )
-  {
-    fprintf(stderr, "sysctl(%s...) expected %lu, got %lu\n",
-                    name, (unsigned long) len, (unsigned long) nlen);
+  if (nlen != len) {
+    fprintf(stderr, "sysctl(%s...) expected %lu, got %lu\n", name, (unsigned long)len,
+            (unsigned long)nlen);
     exit(errno);
   }
   times[CP_IDLE] = cpu_states[4];
@@ -69,7 +62,7 @@ int la_read(unsigned long long *times)
   times[CP_NICE] = cpu_states[1];
   times[CP_USER] = cpu_states[0];
   times[CP_IOWAIT] = cpu_states[3];
-  
+
   return LA_NOERR;
 }
